ADD Bindings

+++ DotnetMVCProject/Models/Employee.cs
@@ -4,11 +4,10 @@ public class Employee
 {
     public int Id { get; set; }
     public string Name { get; set; } = null!;
-    public int Salary { get; set; }
+    public decimal Salary { get; set; }
     public string JopTitle { get; set; } = null!;
     public string? ImageUrl { get; set; }
     public string Address { get; set; } = null!;
     public int? DepartmentId { get; set; }
     public Department? Department { get; set; }
-
 }
